Added localization converters for unknown (= empty) album and artist
This should be a good starting point for having this part of app localized. I tried to include all the shared parts, but focused on Windows Phone only.
This commit is a complement to the commit where the 'Unknown album' and 'Unknown artist' is removed from the code and replaced by string.Empty.
No known issues with this approach so far, however, they might happen.

Signed-off-by: Pavel Valach <valach.pavel@gmail.com>
Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>