Use Dependency Property "trackitem" to display actual info
This is supposed to improve XAML performances. In fact, we don't use
proper bindings which work using reflection, so we save some CPU cycles.
Also, this patch changes the design a bit (with an icon explaining the track
is playing), and obviously fixes some old bugs due to previous poor
performances and unreliable architecture).