Update slideshow logic : based on VMs events
This complete rewrite of the Slideshow engine allows us to quickly update
pictures according to the context, for a better UX.
For example, when navigating in the music library lists, all pictures
are randomly displayed. However, when navigating to an artist/album page,
the related pictures are displayed.
Also, some pages force their theme to Dark, and now the engine listens to
these changes and updates the background accordingly.