Fix crash with hard cast ArtistItem in ArtistHorizontalTemplate
When getting an artist for the ArtistHorizontalTemplate, it does not return an ArtistItem, but an Int. This causes the view to break, causing a crash. We need to figure out why it's doing that but for a quicker fix, remove the hard cast and make it soft. Which should at least let it fail gracefully.