Remove old translation files, replace with shared translation resw
This uses the multilingual app toolkit V4 to support all of our WinRT projects. Including the Windows 8.1, Windows Phone 8.1, and UWP. Now all we need to do to translate the rest of the app is send off the XLF files. Then once we build the app, it will compile the resw's to give the correct language.

And this time it should actually work :)