Fix ui thread marshalling errors on startup for video page
There were thread marshalling errors when starting up the app. It would try to populate the video library on the start screen during app initialization in a different thread from the UI, which caused the video library screen to appear empty. I put adding the media file into the collection on the UI thread, which fixed it. So now the library should populate on first load.