Update SQLite Extension to 3.13
Note: The anniversary update for Windows 10 includes SQlite built in (https://blogs.windows.com/buildingapps/2016/05/03/data-access-in-universal-windows-platform-uwp-apps/) We may want to use that instead of using the extension, since needing to update it every time there is a new release is a pain. It's an older version, but you won't have to install any extra extensions or addons to use it.