Replace Popup with resizing current window for PiP
Making the video show in Picture in Picture in a popup is very hard, due to threading and being able to access the video stream from the original swap chain panel.
However, we can instead resize the current window and turn that into a PiP. Since we already handle removing UI elements for a small window, this makes it an easy way to get PiP working,
since all we need to do is call CompactOverlay on the existing window, and everything else is handled for us.